abstracted connections away from individual service callers
